Add files via upload
This is a python practice problem from Facebook Meta preparation HUB. You are given an input dictionary with keys as strings, and values as integers. Complete a function that returns the key with the nth smallest value. If the solution involves two keys that have the same value, return the key that is lexicographically earliest. If n is larger than the number of distinct keys or equal to 0, then return null.